+++ CHANGELOG.md
@@ -6,6 +6,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 ### Added
+- SDAP-472: Added support for defining Zarr collections in the collection config
 ### Changed
 - Improved Collection Manager logging
   - Inhibited overly verbose loggers